fix(python): implement python icon patching
Also moved files from `resources/icons` to
`resources/jetbrains_icons/icons`.
You might ask why this is necessary. It is necessary because a company
with a yearly revenue of 430 Million US Dollars couldn't be bothered
with implementing proper resource separation between plugins.
So if we leave the icons in the `resources/icons` folder, and another
plugin does the same, it's entirely possible that each plugin
accidentally uses icons of the other plugin.

The next question is why this fix took so long.
To fix this we're relying on APIs that are neither documented in the
official SDK Docs, nor the Javadoc, so we needed to rely on hearsay from
other plugin authors to realize this option even exists.

Signed-off-by: thelooter <evekolb2204@gmail.com>